This Class has three public methods:
getStorageName():  Resolves links in the CS to the target SE name.

getStorage():      This creates a single storage stub based on the parameters passed in a dictionary. This dictionary must have the following keys: 'StorageName', 'ProtocolName', 'Protocol'. Optional keys are 'Port','Host','Path','SpaceToken'.

getStorages()      This takes a DIRAC SE definition and creates storage stubs for the protocols found in the CS. By providing an optional list of protocols it is possible to limit the created stubs. This is called by StorageElement when it is initialized.


git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@5080 4525493e-7705-40b1-a816-d608a930855b