When initialising can pass flag whether or not to use the repository. In addition can pass the location of the repository. If no location is provided then JobRepository() will first try to put it in $HOME/.repo.cfg and if $HOME is not defined then in `pwd`/.repo.cfg.
If the repo is initialised then when jobs are submitted using the submit() they are automatically added to the repository. Also, when the monitor method in used the new status is automatically updated in the repo.

TODO: When the job output is retrived update the repo and if the output data is retrieved the update the repo.


git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@16242 4525493e-7705-40b1-a816-d608a930855b