script to allow users to obtain the files from their catalog directory based on the age of the files and a filemask.
recursively loops over the contents of a supplied directory checking the creation time of the elements found. any subdirectories found which are older than the supplied age they are added to the recursive search. once we reach the leaf directories the creation time of the files are checked and also the wildcard compared with the full LFN. matching files are stored as a list and written out to a local file.


git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@17165 4525493e-7705-40b1-a816-d608a930855b