Add the possibility to take the user/group information directly from the CS.
With this approach there is no need to add/remove users and groups to the catalog since any user that is registered in the CS (and able to contact the service) will be able to use the catalog. Since there will be no additional table to store the user/groups no integer uid/gid will exist. Therefore the user and group information will be stored as strings in the FC_FileInfo and FC_DirectoryInfo tables. 

The performance of an index on a 8 character string vs an integer is not huge, although this should be investigated. Some existing results can be found here:

http://www.mysqlperformanceblog.com/2008/01/24/enum-fields-vs-varchar-vs-int-joined-table-what-is-faster/


git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@22325 4525493e-7705-40b1-a816-d608a930855b