CHANGE: Added more status report
Several common errors where not being reported: missing application, and in case of failure, no status update was made. Added a self.__report if the application is not found, as well as a self.__report if the application failed. In that case, the last 200 characters of the thread output are added as job parameter with the name ApplicationError