CHANGE: Propose recursive treatment in FC change operation
Until now, the recursive options where not used. I propose a way to do
that: if the recursive option is required, first get the directory tree
(that can be slow), and then change the properties for every item. It's
assumed that all the sub directories have the same rights as the parent
as the security check is only done for the top directory/file.