NEW: getCurrentStatus method
For the RSS, the SpaceTokenOccupency should use the StorageElement
instead of the lcg_util calls (to allow monitoring also DIRAC
StorageElements), therefore I added the method to extract the info from
the corresponding methods. 

In any case, the space token is a bit nasty: all SEs do not provide
dedicated ones for the VOs. Catch the case the space token is not
defined, and return an error.