NEW: SSHOSGComputingElement and condorgce script
To submit to OSG, a variation of the condor client must be provided. In
particular, it requires the "grid" Universe, and the "grid_resource"
parameter. The condor_q is also different: the user is not needed as the
submitter is the submission host (apparently, from the doc), but adding
a -globus command line option allows for grid job monitoring.

This is untested as ILCDIRAC does not have an OSG client installed
(would be nice though).