Update pilotCommands.py
This replaces the regular expression '.cfg&' for the extra options file names test (which would match files with "cfg&" anywhere in them) with a test for the suffix ".extra.cfg" and uses the contents of the files rather than the name of the files as the extra options. I've also placed these extra options at the start of the generated options, as some (eg -o CEType) don't work if placed at the end.