Another fix for populating script
The script now checks the setup of the hosts to register into the DB
only installations from the same setup.
Also, the script now checks whether there is a MySQL installation in the
hosts before trying to retrieve their DB installations ( before, trying
to do so when MySQL is not present would bause the host to be ignored,
not registering the rest of its components ).