Several changes to ComponentMonitoring
Due to the fact that the requests to the ComponentMonitoring service are
currently made by the server's SysAdmin service ( and not from the
original caller of the top level script/CLI command ) the certificate of
the server is used instead, meaning that the calls turn out
unauthorized.

To avoid this, the InstallTools' monitorInstallation/Uninstallation
funcions that make the requests to the ComponentMonitoring service have
been moved so that they are callable by the scripts or the CLI.

Also, some minor changes to the InstalledComponentsDB module have been
made to avoid sqlAlchemy warnings.