Fix on CLI
Fixed a bug that made the CLI crash upon calling the 'show
installations' command when there was an entry such that the
UninstallationTime was present, but not the person that unistalled the
component. Now it is not assumed that the uninstalled person is always
registered, even if the component is uninstalled.