PseudoDynamic monitoring
Added a new table to the InstalledComponents database to store the
information that was previously returned by the getHostInfo method
exported by the SysAdmin service.
A periodic task takes care of updating this table with the latest stats
from the host.
Remade getHostInfo call in the SysAdmin service to retrieve this
information from the database instead of calculating it every time.