Requested script to run jobs locally using job().runlocal() from the API
this script gets the VO extension name (e.g. LHCb) in general and uses it to create the job object.
The original script was created to run LHCb jobs, so must of its functions might not be usefull to other VOs,
or might need extra procedures to make it functional for other VOs...

Marcelo Vilaca