fix: Avoid producing massive log output in OptimizationMind
The OptimizationMind suffers from a rare race condition that causes tasks
to get lost occasionally. I thought I'd fixed this in #5667 however it's
much too difficult to fix completely it would require the internal state
to be entirely refactored. Reverting this line avoids massive amounts of
log output when this happens.

We've been running with this as a hotfix in LHCb since Decemeber so it
should be safe.