feat(BDII2CSAgent): allow using AREX or ARC6 CEs instead of ARC
If the ComputingShareEndpoints of an ARC like CE advertises an emies
endpoint, we are going to use AREX as the CE, even if gridftpjob
endpoint is also advertised.  If only gridftpjob is advertised we go
with ARC6. If neither exists, we error out.

Also do not overwrite if we have already set AREX or ARC6 as the CE.