fix: Modified dirac_cert_convert.py to allow pcks12 certificate with legacy crypto
Some providers (cf. CNRS) use certificates with outdated crypto,
so need to call `openssl` with the "-legacy" option.

Modified `dirac_cert_convert.py` to add a switch for this ("--legacy") using the `@Script` mechanisms.

Added the long option only, not putting the short option ("-l") since this should be exceptional.

Also added a gLogger.warn() for this.